web: moved route pages to nested directory format
