+++ apps/web/src/routeTree.gen.ts
@@ -15,11 +15,11 @@ import { Route as MemoriesRouteImport } from './routes/memories'
 import { Route as ConnectorsRouteImport } from './routes/connectors'
 import { Route as AutomationsRouteImport } from './routes/automations'
 import { Route as IndexRouteImport } from './routes/index'
-import { Route as RecordingsIndexRouteImport } from './routes/recordings.index'
+import { Route as RecordingsIndexRouteImport } from './routes/recordings/index'
 import { Route as AutomationsIndexRouteImport } from './routes/automations.index'
 import { Route as AgendaIndexRouteImport } from './routes/agenda.index'
 import { Route as SessionIdRouteImport } from './routes/session.$id'
-import { Route as RecordingsIdRouteImport } from './routes/recordings.$id'
+import { Route as RecordingsIdRouteImport } from './routes/recordings/$id'
 import { Route as AutomationsAutomationIdRouteImport } from './routes/automations.$automationId'
 import { Route as AgendaListIdRouteImport } from './routes/agenda.$listId'
 import { Route as AutomationsSessionsIdRouteImport } from './routes/automations.sessions.$id'
