+++ apps/web/src/routes/recordings/$id.tsx
